update ze_hr_dead_center (#214)
* 0725

·修复错误的道具返还计算

* 0728

·添加双伤脚本
·预特感添加低重力与地面检测

* 0728

·修复严重bug

* 0730

·防止Tank丢石头过程中ztele

* 0803

·新增Spitter脚本
·调整连关发枪时机

+++ 2001/ze_hr_dead_center/save_data.js
@@ -4,7 +4,7 @@ import { CSPlayerPawn, Instance, CSGearSlot } from "cs_script/point_script";
  * 数据存读取脚本
  * 此脚本用于实现回合结算后仍能恢复连关数据
  * 此脚本由皮皮猫233编写
- * 2026/7/16
+ * 2026/8/3
  */
 
 const weaponSlots = [CSGearSlot.RIFLE, CSGearSlot.PISTOL];
@@ -31,13 +31,13 @@ Instance.OnScriptInput("SaveData", () => {
         if (health <= 0) continue;
         const armor = player.GetArmor();
         const weapons = FindWeapons(player);
-        playerData.set(player, { 
-            armor: armor, 
-            health: health, 
-            weapons: weapons, 
-            items: FindItems(player), 
-            helmet: player.HasHelmet(), 
-            money: player.GetPlayerController()?.GetMoneySpendableNow() 
+        playerData.set(player, {
+            armor: armor,
+            health: health,
+            weapons: weapons,
+            items: FindItems(player),
+            helmet: player.HasHelmet(),
+            money: player.GetPlayerController()?.GetMoneySpendableNow()
         });
     }
 });
@@ -53,12 +53,12 @@ Instance.OnScriptInput("ReadData", () => {
         Instance.ServerCommand("c_revive @t");
     });
 
-    // 延迟5秒后读取数据
-    Delay(5, () => {
+    // 延迟2秒后读取数据
+    Delay(2, () => {
         Instance.ServerCommand("say **正在读取连关数据**");
         const players = /** @type {CSPlayerPawn[]} */ (Instance.FindEntitiesByClass("player"));
         for (const player of players) {
-            if (!player || !player.IsValid()) continue;
+            if (!player.IsValid()) continue;
 
             // 检查是否被存储为上局存活的玩家
             if (playerData.has(player)) {
@@ -75,19 +75,22 @@ Instance.OnScriptInput("ReadData", () => {
                     }
                 }
             } else {
-                const playerController = player.GetPlayerController();
-                if (playerController && playerController.IsValid()) {
-                    Instance.ServerCommand("c_infect #" + playerController.GetPlayerSlot());
-                    player.Teleport({ position: { x: 0, y: 0, z: -3560 } });
-                    Delay(0.1, () => player.Teleport({ position: { x: 0, y: 0, z: -3560 } }));
-                    Delay(0.2, () => player.Teleport({ position: { x: 0, y: 0, z: -3560 } }));
-                    Delay(0.3, () => player.Teleport({ position: { x: 0, y: 0, z: -3560 } }));
+                Delay(5, () => {
+                    if (!player.IsValid()) return;
+                    const playerController = player.GetPlayerController();
+                    if (playerController && playerController.IsValid()) {
+                        Instance.ServerCommand("c_infect #" + playerController.GetPlayerSlot());
+                        player.Teleport({ position: { x: 0, y: 0, z: -3560 } });
+                        Delay(0.1, () => player.Teleport({ position: { x: 0, y: 0, z: -3560 } }));
+                        Delay(0.2, () => player.Teleport({ position: { x: 0, y: 0, z: -3560 } }));
+                        Delay(0.3, () => player.Teleport({ position: { x: 0, y: 0, z: -3560 } }));
 
-                    // 延迟1秒后检测是否仍为人类防止部分社区禁止c_infect指令
-                    Delay(1, () => {
-                        if (player.IsValid() && player.GetTeamNumber() === 3) player.Kill();
-                    });
-                }
+                        // 延迟1秒后检测是否仍为人类防止部分社区禁止c_infect指令
+                        Delay(1, () => {
+                            if (player.IsValid() && player.GetTeamNumber() === 3) player.Kill();
+                        });
+                    }
+                });
             }
         }
         playerData.clear();
@@ -172,17 +175,17 @@ function FindItems(player) {
  * @param {number[]} items
  */
 function GiveWeaponsAndItems(player, weapons, items) {
-    for (const weapon of weapons) {
-        player.DestroyWeapons();
-        Delay(1, () => {
+    player.DestroyWeapons();
+    Delay(1, () => {
+        for (const weapon of weapons) {
             player.GiveNamedItem(weapon);
-            for (let i = 0; i < itemNames.length; i++) {
-                for (let number = 0; number < items[i]; number++) {
-                    player.GiveNamedItem(itemNames[i]);
-                }
-            }
-        });
-    }
+        }
+        for (let i = 0; i < itemNames.length; i++) {
+            if (items[i] <= 0) continue;
+            player.GiveNamedItem(itemNames[i]);
+            player.FindWeapon(itemNames[i])?.SetReserveAmmo(items[i]);
+        }
+    });
 }
 
 /** @type {{ time: number, callback: () => void }[]} */
@@ -238,7 +241,7 @@ function QueueThink(time, callback) {
  */
 function RunThinkQueue() {
     const currentTime = Instance.GetGameTime();
-    
+
     // 执行所有到期的任务
     while (thinkQueue.length > 0 && thinkQueue[0].time <= currentTime) {
         const task = thinkQueue.shift();
